fix: Spring 7 - reinstate suffix/trailing-slash matching in the handler mapping (literal-first)
Replace the blind MediaTypeSuffixFilter with literal-first matching in
CustomRequestMappingHandlerMapping.getHandlerInternal: try the path as-is (so
controllers that map an extension literally, e.g. OpenApiController's
/openapi/openapi.json, keep working), then fall back to matching the path with a
trailing slash and/or a registered media-type extension removed (recording the media
type for SuffixMediaTypeContentNegotiationStrategy). Fixes OpenApiControllerTest (10/10)
and trailing-slash routing without a separate filter. Removed MediaTypeSuffixFilter +
its registrations.

Known residual: 2 DataSetControllerTest metadata-download tests (.json/.json.zip
Content-Disposition filename) still fail - the export converter's format/filename
derivation needs the path extension; tracked as follow-up.
