+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/mvc/CustomRequestMappingHandlerMapping.java
@@ -29,8 +29,15 @@
  */
 package org.hisp.dhis.webapi.mvc;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequestWrapper;
 import java.lang.reflect.Method;
+import java.util.Map;
+import org.hisp.dhis.webapi.security.config.WebMvcConfig;
+import org.hisp.dhis.webapi.view.SuffixMediaTypeContentNegotiationStrategy;
+import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.mvc.condition.RequestMethodsRequestCondition;
 import org.springframework.web.servlet.mvc.method.RequestMappingInfo;
 import org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandlerMapping;
@@ -47,8 +54,23 @@ import org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandl
  * 43. It also still allows the endpoint without any version e.g. <br>
  * <code>/api/icons/{key}</code> <br>
  * which is what we want clients to use exclusively in the future.
+ *
+ * <p>It also reinstates the path-extension and trailing-slash matching that Spring Framework 7.0
+ * removed from the handler mapping (the {@code setUseSuffixPatternMatch} / {@code
+ * setUseTrailingSlashMatch} flags). For each request the canonical (literal) path is matched first
+ * — so controllers that map an extension literally (e.g. {@code OpenApiController}'s {@code
+ * /openapi/openapi.json}) keep working — and only if that has no handler do we fall back to
+ * matching the path with a trailing slash and/or a registered media-type extension removed (e.g.
+ * {@code /api/dataElements.json} → {@code /api/dataElements}), recording the resolved media type
+ * for {@link SuffixMediaTypeContentNegotiationStrategy}.
  */
 public class CustomRequestMappingHandlerMapping extends RequestMappingHandlerMapping {
+  private static final char EXTENSION_SEPARATOR = '.';
+
+  private static final char PATH_SEPARATOR = '/';
+
+  private final Map<String, MediaType> mediaTypes = WebMvcConfig.MEDIA_TYPE_MAP;
+
   @Override
   protected RequestMappingInfo getMappingForMethod(Method method, Class<?> handlerType) {
     RequestMappingInfo info = super.getMappingForMethod(method, handlerType);
@@ -57,9 +79,7 @@ public class CustomRequestMappingHandlerMapping extends RequestMappingHandlerMap
       return null;
     }
 
-    // Spring 7.0 removed the suffix-pattern / trailing-slash flags from PatternsRequestCondition;
-    // path-extension content negotiation is now handled by MediaTypeSuffixFilter. The only mapping
-    // adjustment that remains is to default a method-less mapping to GET (legacy DHIS2 behaviour).
+    // Default a method-less mapping to GET (legacy DHIS2 behaviour).
     RequestMethodsRequestCondition methodsCondition = info.getMethodsCondition();
 
     if (methodsCondition.getMethods().isEmpty()) {
@@ -67,5 +87,97 @@ public class CustomRequestMappingHandlerMapping extends RequestMappingHandlerMap
     }
 
     return info;
+  }
+
+  @Override
+  protected HandlerMethod getHandlerInternal(HttpServletRequest request) throws Exception {
+    // 1) Match the path as-is so literal-extension mappings (e.g. /openapi/openapi.json) win.
+    HandlerMethod handlerMethod = super.getHandlerInternal(request);
+    if (handlerMethod != null) {
+      return handlerMethod;
+    }
+    // 2) Fall back to a normalised path (trailing slash and/or registered extension removed).
+    HttpServletRequest normalized = normalize(request);
+    return normalized != request ? super.getHandlerInternal(normalized) : null;
+  }
+
+  /**
+   * Returns a request exposing the path with a single trailing slash and/or a trailing registered
+   * media-type extension removed (recording the extension's media type for content negotiation), or
+   * the original request when there is nothing to normalise.
+   */
+  private HttpServletRequest normalize(HttpServletRequest request) {
+    String uri = request.getRequestURI();
+    boolean trailingSlash = uri != null && uri.length() > 1 && uri.endsWith("/");
+    String pathForExtension = trailingSlash ? uri.substring(0, uri.length() - 1) : uri;
+    String extension = getRegisteredExtension(pathForExtension);
+
+    if (!trailingSlash && extension == null) {
+      return request;
+    }
+    if (extension != null) {
+      request.setAttribute(
+          SuffixMediaTypeContentNegotiationStrategy.SUFFIX_MEDIA_TYPE_ATTRIBUTE,
+          mediaTypes.get(extension));
+    }
+    return new PathNormalizingRequestWrapper(request, extension, trailingSlash);
+  }
+
+  /**
+   * Returns the registered extension (a key of {@link #mediaTypes}) at the end of the given URI's
+   * last path segment (everything after the first dot, e.g. {@code adx.xml.gz}), or {@code null}.
+   */
+  private String getRegisteredExtension(String uri) {
+    if (uri == null) {
+      return null;
+    }
+    int lastSeparator = uri.lastIndexOf(PATH_SEPARATOR);
+    int dotIndex = uri.indexOf(EXTENSION_SEPARATOR, lastSeparator + 1);
+    if (dotIndex == -1) {
+      return null;
+    }
+    String extension = uri.substring(dotIndex + 1);
+    return mediaTypes.containsKey(extension) ? extension : null;
+  }
+
+  /** Wraps a request to expose the path with a trailing slash and/or {@code .extension} removed. */
+  private static final class PathNormalizingRequestWrapper extends HttpServletRequestWrapper {
+    private final String suffix;
+    private final boolean trailingSlash;
+
+    PathNormalizingRequestWrapper(
+        HttpServletRequest request, String extension, boolean trailingSlash) {
+      super(request);
+      this.suffix = extension != null ? EXTENSION_SEPARATOR + extension : null;
+      this.trailingSlash = trailingSlash;
+    }
+
+    private String strip(String path) {
+      if (path == null) {
+        return null;
+      }
+      if (trailingSlash && path.length() > 1 && path.charAt(path.length() - 1) == PATH_SEPARATOR) {
+        path = path.substring(0, path.length() - 1);
+      }
+      if (suffix != null && path.endsWith(suffix)) {
+        path = path.substring(0, path.length() - suffix.length());
+      }
+      return path;
+    }
+
+    @Override
+    public String getRequestURI() {
+      return strip(super.getRequestURI());
+    }
+
+    @Override
+    public String getServletPath() {
+      return strip(super.getServletPath());
+    }
+
+    @Override
+    public String getPathInfo() {
+      return strip(super.getPathInfo());
+    }
   }
 }
